Crypto_project/requirements.txt
Django>=4.0,<5.0
psycopg2  
requests
pycoingecko
python-telegram-bot
scikit-learn
pandas
numpy
chart.js
